Elevated: Fix more size_t confusion

+++ games/elevated/stats/PowerStatsListener.cpp
@@ -15,7 +15,7 @@ void PowerStatsListener::on_elevator_moved(Time, Height height_travelled, Height
 
 size_t PowerStatsListener::total_distance_travelled() const
 {
-    return std::accumulate(m_distance_moved.begin(), m_distance_moved.end(), 0ul, [](uint64_t acc, auto& entry) {
+    return std::accumulate(m_distance_moved.begin(), m_distance_moved.end(), size_t(0), [](uint64_t acc, auto& entry) {
         return acc + entry.second;
     });
 }
